+++ src/init.c
@@ -127,6 +127,9 @@ extern SEXP vctrs_df_list(SEXP, SEXP, SEXP);
 extern SEXP vctrs_identify_runs(SEXP);
 extern SEXP vctrs_locate_runs(SEXP, SEXP);
 extern SEXP vctrs_detect_runs(SEXP, SEXP);
+extern SEXP vctrs_df_slice_complete(SEXP);
+extern SEXP vctrs_df_locate_complete(SEXP);
+extern SEXP vctrs_df_detect_complete(SEXP);
 
 
 // Maturing
@@ -271,6 +274,9 @@ static const R_CallMethodDef CallEntries[] = {
   {"vctrs_identify_runs",              (DL_FUNC) &vctrs_identify_runs, 1},
   {"vctrs_locate_runs",                (DL_FUNC) &vctrs_locate_runs, 2},
   {"vctrs_detect_runs",                (DL_FUNC) &vctrs_detect_runs, 2},
+  {"vctrs_df_slice_complete",          (DL_FUNC) &vctrs_df_slice_complete, 1},
+  {"vctrs_df_locate_complete",         (DL_FUNC) &vctrs_df_locate_complete, 1},
+  {"vctrs_df_detect_complete",         (DL_FUNC) &vctrs_df_detect_complete, 1},
   {NULL, NULL, 0}
 };